dataform

+++ demo/progressor_main.cpp
@@ -7,7 +7,6 @@
 
 using namespace dogm;
 
-// 점유 확률을 하나의 값으로 변환하는 함수
 float pignistic(const GridCell& cell) {
     return cell.occ_mass + 0.5f * (1.0f - cell.occ_mass - cell.free_mass);
 }
@@ -43,7 +42,7 @@ int main(int argc, char** argv) {
 
     while(loader.hasNextFrame()) {
         SensorFrame frame = loader.getNextFrame();
-        float dt = (last_timestamp < 0) ? 0.1 : (frame.timestamp - last_timestamp);
+        float dt = (last_timestamp < 0) ? 0.1f : static_cast<float>(frame.timestamp - last_timestamp);
         last_timestamp = frame.timestamp;
 
         auto start = std::chrono::high_resolution_clock::now();
@@ -61,7 +60,7 @@ int main(int argc, char** argv) {
             for (int x = 0; x < grid_size; ++x) {
                 const auto& cell = grid_cells[y * grid_size + x];
                 float prob = pignistic(cell);
-                if (prob > 0.15f && prob < 0.85f) { // 유의미한 데이터만 저장
+                if (prob > 0.15f && prob < 0.85f) {
                      output_file << std::fixed << std::setprecision(4) << frame.timestamp << ","
                                  << x << "," << y << "," << prob << ","
                                  << cell.mean_x_vel << "," << cell.mean_y_vel << "\n";
